refactor(model): replace createModelInstance factory helper with static define method
