+++ test/model/Model.test.ts
@@ -1,28 +1,37 @@
-import type { ModelAttrs } from '@src/model/BaseModel';
-import Model from '@src/model/Model';
+import { DbCollection } from '@src/db';
+import { Model, type ModelAttrs } from '@src/model';
 
 interface UserAttrs extends ModelAttrs<number> {
   name: string;
   email: string;
 }
 
+const UserModel = Model.define<UserAttrs>();
+
 describe('Model', () => {
-  describe('create', () => {
-    it('should create a model instance with attribute getters/setters', () => {
-      const user = Model.create<UserAttrs>({
+  let collection: DbCollection<number, UserAttrs>;
+
+  beforeEach(() => {
+    collection = new DbCollection<number, UserAttrs>({ name: 'users' });
+  });
+
+  describe('define', () => {
+    it('should create a model class with attribute getters/setters', () => {
+      const user = new UserModel({
         name: 'User',
         attrs: { name: 'John', email: 'john@example.com' },
+        collection,
       });
-
       expect(user.name).toBe('John');
       expect(user.email).toBe('john@example.com');
       expect(user.isNew()).toBe(true);
     });
 
     it('should allow attribute modification', () => {
-      const user = Model.create<UserAttrs>({
+      const user = new UserModel({
         name: 'User',
         attrs: { name: 'John', email: 'john@example.com' },
+        collection,
       });
 
       user.name = 'Jane';
@@ -31,9 +40,10 @@ describe('Model', () => {
     });
 
     it('should preserve model methods', () => {
-      const user = Model.create<UserAttrs>({
+      const user = new UserModel({
         name: 'User',
         attrs: { name: 'John', email: 'john@example.com' },
+        collection,
       });
 
       expect(typeof user.save).toBe('function');
